app/requirements.txt
Flask
Flask-SQLAlchemy
SQLAlchemy
PyMySQL
pytest
cryptography
python-dotenv